move pricing context field config to new file
